[V8] Use correct swarming priority on V8 tryserver
Previously, V8 had a code path to allow manual tryjobs with higher priority
than CQ. After the luci migration, the path became dead code and all CQ
tasks were run with too high prio.

This removes the old logic and uses the same priority as in Chromium for
all tryjobs. The requester field seems to be not set anymore but seems to
also not be needed for anything. Neither is the purpose tag. The user
property is None in production tasks, so it's removed here, too.

TBR=dpranke@chromium.org

Bug: 876351
Change-Id: I82f734ae73c8a9bda1d9ee2ad13786fa6d33e7f7
Reviewed-on: https://chromium-review.googlesource.com/1186586
Reviewed-by: John Budorick <jbudorick@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>