[buildbucket] Update existing lease key in poll
When leasing a build in the poll loop, if we already have a stale lease for the
same build id, update its lease key instead of scheduling a new build.

This shifts primary lease key storage to the in-memory _leases dict, as opposed
to "lease_key" available in build properties, and allows them to be different.
Note that _leases is not persisted, so on master restarts we may load
non-updated lease keys. It happens only on master restart.

This CL does not update the send_heartbeats code path, which is another place
where we may find out that a lease key is stale. Added a TODO.

Bug: 877388
Change-Id: I5c0ce484d15d60e535b7b896108d155272b035ea
Reviewed-on: https://chromium-review.googlesource.com/1188628
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>