Add config for chromium.webrtc.fyi.experimental.
The idea is to duplicate the config for chromium.webrtc.fyi
everywhere so I can safely experiment with my new test suites and
make sure the config is in the right place. It's possible that there
is unnecessary config in the new SPEC; I just generated it out of the
old one. The only modification is that I removed tests, which will be
specified by waterfalls.pyl.

I also need CLs for mb_config.pyl and maybe test_results as well.

Bug: 877018
Change-Id: I46f6735dcd10ca299ea60a32029a79ecf181ad07
Reviewed-on: https://chromium-review.googlesource.com/1201588
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Patrik Höglund <phoglund@chromium.org>