Report the testers being run by a trybot.
A BotId namedtuple type was added to BotConfig to encapsulate the logic
of computing the tester mastername when a tester is part of the bot id.
Additional methods were added to BotConfig to provide read-only access
to the bot ids and to be able to get the bot type for a given bot id.

Bug: 866624
Change-Id: I80db8e84a8f9bc47f30c9e8c2233cc934e3d46d0
Recipe-Nontrivial-Roll: build_limited_scripts_slave
Reviewed-on: https://chromium-review.googlesource.com/1222386
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>