Remove the method test_utils.determine_new_failures.
This CL is a refactor with no intended behavior change.

The method determine_new_failures is used in exactly two places -- the
chromium_tests recipe module and the blink_downstream recipe.

There are three problems with determine_new_failures.
  * The method takes a closure that references global state. Running this
    closure has significant side effects. This makes it difficult to follow the
    control flow.
  * The retry logic for chromium_tests will be changing in the future. It will
    be difficult to modify determine_new_failures to take on the new behavior,
    partly due to the closure mentioned above.
  * The retry logic for chromium_tests will be diverging from that of
    blink_downstream.

This CL removes the method. The logic was copied into blink_downstream with no
changes. The functionality in chromium_tests is identical to before, but the new
logic is significantly simpler.

This CL removes some debugging logic added in
https://chromium-review.googlesource.com/c/chromium/tools/build/+/1115874 that
is no longer necessary.

Bug: 883321, 850653
Change-Id: Ie575ddcff824007a7543c344b327e1d7d86bb449
Reviewed-on: https://chromium-review.googlesource.com/1222648
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>