Add support for 'retry with patch' to chromium_tests base Test class.
This CL does not change any production functionality. The new logic is only used
in tests.

This CL adds a method tests_to_retry() to the Test base class. It can be called
from any subclass and will correctly return the tests to run.

This allows removal of quite a bit of duplicated code. It also allows for the
addition of a new suffix 'retry with patch'.

This CL fixes one small bug in order to make tests pass:
  * In blink_test.py, the 'without patch' step was never being run because it
    was outside of the finally block.

A side effect of this CL is that test suites will never be passed an empty list
of tests to filter/test. This behavior is poorly defined [should all tests run?
no tests run?].

Change-Id: I0ec82b0ca4fdaf597b1ef713a8eba10a37c3ab5a
Bug:883321
Reviewed-on: https://chromium-review.googlesource.com/1225712
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>