Collect last test if it's the only test still running
When collecting a list of tasks in test_utils, and there's only one test
still running, just skip straight to collecting it. The current
swarming logic polls the swarming server once a minute, but if there's
only one test we're waiting for, just call the normal logic for that
test. No need to poll the server.

Bug: 828937
Change-Id: Id4462f4da124d8741fb149691e6be5fc19bb2bcb
Reviewed-on: https://chromium-review.googlesource.com/1232135
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>