Always pass step name when archiving layout test results
It turns out that the current layout test archival links have no concept
of if they're with patch, without patch, etc.. This CL makes every archive
that the archival step generates explicitly say if it has a patch or not.
This will allow for uploading of layout test results with the
'without patch' suffix, without breaking the existing links.

Bug:882946
Change-Id: Idd07d5eeaad6f810b93e22c79dde945d8d374645
Reviewed-on: https://chromium-review.googlesource.com/1232713
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>