binary_size_trybot.py: Add additional checks and results.
This CL changes the way the helper trybot script is used. Now, the
helper script is given diff output paths and is in charge of creating
all diffs. The recipe simply exposes these output files in separate
steps, and checks one of the output paths for expected text to
determine whether or not an undocumented size increase exists.

Additionally, the bot now shows added/removed dex method names and
includes a check for uncompressed pak files and big jumps in the dex
method count.

Lastly, the recipe also includes a step that archives the .ndjson
report created by the trybot helper script and includes a link to the
report.

Bug: 702625
Change-Id: Ic855213224f2c122d752756d7394c5f9419942ed
Reviewed-on: https://chromium-review.googlesource.com/1239256
Commit-Queue: Eric Stevenson <estevenson@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Reviewed-by: agrieve <agrieve@chromium.org>