Enforce that tester and builder configs are synced.
Because the config objects are essentially global data, in order for
recipes to be able to reason about a builder's config when working
with a tester, the configs must be the same. For example, this is
necessary so that the GN args for a builder can be looked up when
running a tester.

Bug: 732269
Change-Id: I6815897d895eb57ddb42830e612efe76bd68f35f
Reviewed-on: https://chromium-review.googlesource.com/1235295
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>