Enable overriding runtests values in chromium.c.runtests.
https://crrev.com/c/1235295 enforces that a tester's settings for
chromium configs match its builder's so that code can reason about the
tester's builder. Currently one of the builders is not synced because
its builder builds with LSan but it doesn't run with LSan and this is
controlled by the chromium config. More generally, setting tags for
swarming when running tests is also affected by some of the chromium
configuration values. Providing the ability to override this using
chromium_tests' configuration enables keeping the chromium configuration
synced between builder and tester and still change the observed
configuration when running the tester.

Bug: 732269
Change-Id: Ib149f0703d202f876e91ede2debb6f536d9489d5
Reviewed-on: https://chromium-review.googlesource.com/1246992
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>