Use both sources of properties for chromium_trybot.py
This allows for the modules this recipe depends on to switch to
using the new buildbucket APIs without touching this recipe. The
expectations for the recipe itself may change, but the test data won't
change, which will give us more confidence that the change
is a no-op.

Change-Id: I4010a2790f25b5cea1423d8af753b2a2ccd1b9dc
Reviewed-on: https://chromium-review.googlesource.com/c/1258305
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>