Refactor chromium recipe test properties
No functional changes; this is a refactor to help with migrating
this recipe to use the new api.buildbucket.ci_build test api.
This allows for the modules this recipe depends on to switch to
using the new buildbucket APIs without touching this recipe. The
expectations for the recipe itself may change, but the test data won't
change, which will give us more confidence that the change
is a no-op.

Bug: 889661
Change-Id: I67a8c77315aa334077c0dbb66aa8d094fd7f34a6
Reviewed-on: https://chromium-review.googlesource.com/c/1258319
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>