Relax the requirement that builder and tester configs are synced.
The code for looking up tester GN args has been implemented in a way
that doesn't require the builder and tester configs to be synced to
produce the correct results. Allowing the builder and tester to
deviate enables removing less than desirable code that was used for
disabling LSan when running the "Linux ASan Tests (sandboxed)"
tester. It is still generally desirable that the builder and tester
are synced, so the sync test has been relaxed to enable custom checks.

Change-Id: I562c77d205755c80c5df68a21f90064dc52caaeb
Reviewed-on: https://chromium-review.googlesource.com/c/1265736
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>