Better flake detection with 'retry without patch'.
Previously, 'retry without patch' would look for a single test success and
consider the test as passing on tip of tree. This CL makes it so that 'retry
without patch' looks for a single failure to consider the test flaky on tip of
tree.

TestResult and GTestResult already support pass_fail_counts, a dictionary that
provides the number of failures and successes per test. This CL plumbs that
dictionary through Test subclasses to make it accessible to the CQ recipe.

Future CLs will update the parameters passed to swarming tasks so that tests are
run multiple times, even on success to look for flaky failures.

Bug: 892307
Change-Id: I986d4e1cead1a1275174bd995b4535c887a6c970
Reviewed-on: https://chromium-review.googlesource.com/c/1265655
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Auto-Submit: Erik Chen <erikchen@chromium.org>