Repeat gtests 10 times in the 'retry without patch' step.
In 'retry without patch', we want to see if the test is flaky on tip of tree. To
do so, we must repeatedly run the test, even if it succeeds.

If the test is flaky on tip of tree, then the failure is likely caused by tree
flakiness rather than a problem with the CL.

Bug: 892307
Change-Id: I66c687d68320017e103299128bcb7d07121fcec2
Reviewed-on: https://chromium-review.googlesource.com/c/1269016
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>