Repeat webkit_layout_tests 10 times in the 'retry without patch' step.
In 'retry without patch', we want to see if the test is flaky on tip of tree. To
do so, we must repeatedly run the test, even if it succeeds.

If the test is flaky on tip of tree, then the failure is likely caused by tree
flakiness rather than a problem with the CL.

Bug: 892307
Change-Id: I57264ca6e1473026a62fa2d39155d4ff77a38764
Reviewed-on: https://chromium-review.googlesource.com/c/1271636
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>