[code coverage] Rebase coverage metadata
This CL only rebases and dumps the flat format because it's much easier
to implement and debug as of now for the purpose of prototyping.

Eventually, the flat format should be deprecated in favor of the
compressed format and protobuf should be used explicitly so that the
formats are maintainable and tractable.

Bug: 902404
Change-Id: I2612f5ef2953c8f3ba01e278ac3e0a784c3eaa9b
Reviewed-on: https://chromium-review.googlesource.com/c/1330926
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>