[Dart] Remove --append-logs argument to test.py, clear debug log manually.
This depends on a Dart SDK change where the --append-logs argument is dropped
from test.py, and the only log it applies to, .debug.log, appends always.
The debug log is currently broken in the recipe, so this can be committed
without regressing behavior. Making append the default makes the recipe simpler.

Change-Id: I9bf08afd27b9415ada7e185fdee81d5fabb5c1f1
Reviewed-on: https://chromium-review.googlesource.com/c/1350858
Commit-Queue: William Hesse <whesse@google.com>
Reviewed-by: Martin Kustermann <kustermann@google.com>