Change the Mojo chromium.fyi bots to run swarming tasks at pri=25.
Currently the Mojo bots on chromium.fyi run tasks at the default
swarming priority for .fyi bots (35); this means that they are
lower priority than main waterfall bots (25) and optional CQ
bots (30), and may get starved before the others. We don't want that,
we want them to have the same priority as the main waterfall bots.

Longer-term, we should move them to a new "master" so that we don't
have multiple builders on the same master with different default
priorities, but this is a simple short term fix that doesn't require
coordinating changes across multiple repos.

Bug: 876570
Change-Id: Iac34f03f8fff2f9201529dda8943f58ae7eead46
Reviewed-on: https://chromium-review.googlesource.com/c/1184381
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>