Run Dawn CQ with ToT Dawn.
Without this Dawn's CQ GPU builders try to apply the patch being
tested on the version of Dawn that's in Chromium's DEPS instead of
ToT Dawn, causing failures when patches don't apply.

This will cause all these builders to use ToT Dawn:
 - Dawn's CQ GPU builders (the reason why this is needed)
 - The GPU FYI waterfall
 - ANGLE's CQ GPU builders
 - GPU optional try builders

This is because all of the above are either the GPU FYI builders
or mirror them so they must all have the same config. It involves
some risk that Dawn will break compilation on one of the extra
builders but it is tiny because Dawn's CQ will be testing the same
configuration.

Bug: chromium:901830
Change-Id: Icc48e02ae1f83b9fd846438a250bd09fe5ab3b42
Reviewed-on: https://chromium-review.googlesource.com/c/1350854
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>