[V8] Add multiple swarming shards to flako
This makes flako spawn multiple swarming shards in each attempt.
During calibration, first the swarming shards and then the repetition
multiplier is increased.

This is not using the shard multiplication from the swarming recipe
module until the V8 side is compatible with it (see TODO in code).

The test simulation for flaky tests is simplified by adding a
default for no failures. Only runs with failures are explicitly
mentioned in the test data.

In a follow up, we'll introduce more step nesting and attempt to
reduce the expectation file size again with better filters.

Bug: v8:8251
Change-Id: I7fae771aa8c80f1aae69d73165ef5261df89c119
Reviewed-on: https://chromium-review.googlesource.com/c/1349609
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Sergiy Belozorov <sergiyb@chromium.org>