'retry with patch' should only rerun failing tests 10 times.
'retry with patch' uses --gtest_repeat=10 to rerun failing tests 10 times.
Unfortunately, the default value of --test-launcher-retry-limit=4 causes
deterministically failing tests to be rerun 40 times. This is not the intention.

This CL updates the recipe logic so that if we set --gtest_repeat=10, then we
also set --test-launcher-retry-limit=1.

Bug: 909834
Change-Id: I80b7972d7a08d161b8fa2a1f2c9e3f6c3af6a1a5
Reviewed-on: https://chromium-review.googlesource.com/c/1351695
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>