Fix edge case where --gtest_repeat is set without --gtest_filter.
In rare circumstances, it's possible for test results to be valid, and all tests
to pass, but for the test suite to be considered a failure due to an infra
issue. This can cause the list of failed tests to be empty. In this case,
retries will rerun all tests. We should avoid setting --gtest_repeat.

Bug: 910706
Change-Id: If81cb64bbb391a9883b57b1a465d0d0b708fc926
Reviewed-on: https://chromium-review.googlesource.com/c/1357281
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>