Add a "hard timeout time" parameter to iOS tests in api.py.
This lets tests specify a hard timeout time that is not 1hr.
Autofill tests will be running for ~1hr at the end of Q4, so I plan on
using this to extend autofill's timeout to 3hrs for safety.

In the long term (Q1 2019), I will be sharding autofill automation to
reduce the length of each shard.

Change-Id: If06c0c375aca9455662bec153395f3ed9f615393
Reviewed-on: https://chromium-review.googlesource.com/c/1361801
Commit-Queue: ericale <ericale@chromium.org>
Reviewed-by: Sergey Berezin <sergeyberezin@chromium.org>