[dart] Exit cleanly even if test.py finds test failures.
In the new workflow, we want to know if test.py itself failed, but we don't
want the test.py step to fail if any tests failed. Instead test failures
will cause the builder to fail at a later step if they're unapproved.

Instead we use the test.py --clean-exit option to cause the step to not fail
due to the test results, but only fail if it failed to produce the desired
test results.

Bug: https://github.com/dart-lang/sdk/issues/35200
Change-Id: I90ddc16fb786330f2e7f0b53e943d99fff255f4e
Reviewed-on: https://chromium-review.googlesource.com/c/1370052
Reviewed-by: William Hesse <whesse@google.com>
Commit-Queue: Jonas Termansen <sortie@google.com>