[V8] Fix blink_downstream syncing to pinned revision
The bot_update-on-failure step should sync V8 to the pinned revision inside
Chromium. This has been achieved by not passing a specific --revision
parameter to bot_update.

Recently, additional logic in bot_update always passes the project revision
to the update call (which here is V8). We now make use of a flag to
suppress this behavior.

Bug: 909615
Change-Id: Icb2fac3331584ffc785a4b1a28b7b30b154dfea2
Reviewed-on: https://chromium-review.googlesource.com/c/1367785
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Sergiy Belozorov <sergiyb@chromium.org>