[code coverage] Handle add/delete/rename files and write tests for them
This CL first fixes a bug for rebasing diff when a file is deleted,
before this CL, line in the diff '+++ /dev/null' is treated as a added
line: '+ /dev/null', which is wrong, and this CL fixes it.

This CL also writes tests for rebassing diff of added/deleted/renamed
files.

Change-Id: I6e6af33e4e00179bde4d6c185ed097fa14c4688d
Reviewed-on: https://chromium-review.googlesource.com/c/1370056
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Roberto Carrillo <robertocn@chromium.org>