[Dart] Upload builder's flaky.json to another path for easy data transfer
A tool we use to upload data cannot follow the link to the latest build,
and runs best if all the files are in a subtree and have a distinct name.
Putting the current flaky.json from every builder in an easily accessible
location.

Change-Id: I0d315de1a85829fb611d5791bc2da32746ddd184
Reviewed-on: https://chromium-review.googlesource.com/c/1386852
Auto-Submit: William Hesse <whesse@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>
Reviewed-by: Martin Kustermann <kustermann@google.com>