Add an API for an early check for missing isolated swarm_hashes
This is the rough equivalent to "find_isolated_tests" but for Swarming.
It accepts a list of targets and asserts that they are present in the 'swarm_hashes' property.
And in test mode it populates isolated_tests so that the hashes don't need to be specified separately through test_api.

This also updates an incorrect usage that WebRTC had (find_isolated_tests step always failed because it's not made for swarming, although it happened to make the tests pass).

Bug: chromium:908001
Change-Id: I38117cbd37dbe51857626bb38d4c1d0d506c33c0
Reviewed-on: https://chromium-review.googlesource.com/c/1388505
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Oleh Prypin <oprypin@chromium.org>