Update iOS builders to use same dir for generation and analysis.
Currently, filter.analyze is run against the directory determined by
the chromium config object which is not the same as the directory that
the build files are generated to because the ios recipe module
specifies the directory when calling chromium.mb_gen. While this
doesn't cause any technical issues, it is confusing.

The filter.analyze method has been updated to accept the build output
directory as an argument and the ios module has been updated to pass
the same value to both chromium.mb_gen and filter.analyze.

Bug: 670521
Change-Id: I79429d4e7c71d635401673ae59bdabfafbc901a1
Reviewed-on: https://chromium-review.googlesource.com/c/1392599
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>