Add info on test execution to gtest log
GTest output includes the stdout of the test execution. We currently
display this via a link in the step in the build, but it doesn't really
show how the step ended. This is important if the test timed out,
failed, etc... This CL makes the test display this information now.

Bug: 912672
Change-Id: I3bd5b02ba1534cb794ad212f2a4965d5e5c714cc
Reviewed-on: https://chromium-review.googlesource.com/c/1368497
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>