Use more accurate determination for maximum command line length.
Logic was added to SwarmingIsolatedScriptTest to prevent more than 100 tests
from being passed as a test filter, as this would sometimes cause the maximum
command line character limit to be exceeded on Windows.

A recent refactor unintentionally extended this logic to SwarmingGtestTest. This
was causing test filters to be skipped on non-Windows systems.

This CL changes the logic to instead skip the test filter only if it would
exceed reasonable lengths [7000 on Windows, 90000 on other platforms].

Bug: 918953
Change-Id: I63f90995ea8a45d17c9b9443c357edd7d05e3047
Reviewed-on: https://chromium-review.googlesource.com/c/1396041
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>