[code coverage] Work around fuzzer invalid test results
There on more than 300 fuzzer targets and all of them are failing due
to invalid test results, and the 300+ red steps cause the whole build
to fail as a Infra failure, which is not good because it masks the real
failed steps.

This CL works the issue around with a temporary solution.

Bug: 921300
Change-Id: I78d7c9cb0c7550fe6c70f42f50cd1512a3fb1c36
Reviewed-on: https://chromium-review.googlesource.com/c/1409704
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>