Make Build Recipes Tester use non-experimental led task.
This builder currently is a blocker for the build repo CQ. However, it
runs the led tasks in experimental mode, which, in turn, causes them
to run their test shards with priority 40.

In the 'fast' version of this recipe it takes 40-60 minutes to run during
the day, but ~10 minutes otherwise. Most of this time is pending times
for test tasks. The 'slow' version of this recipe (i.e. any CL which touches
any recipe code in build.git) takes even longer.

The downside of running these in non-experimental mode is that they'll
upload some data to the test results server. However we expect this to be
an acceptable tradeoff vs. blocking the build CQ for very long periods of
time.

R=bpastene@chromium.org, jbudorick@chromium.org, martiniss@chromium.org

Change-Id: I15213f24cb5fd04810c4d442c00fa84414a858e2
Reviewed-on: https://chromium-review.googlesource.com/c/1415847
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>