Update output directory used during analysis for iOS builders.
Currently, filter.analyze is run against the directory determined by
the chromium config object, which is not the same as the directory
that the build files are generated to because the ios recipe module
specifies the directory when calling chromium.mb_gen. While this
doesn't cause any technical issues, it is confusing because the
directory name determined by the chromium config doesn't necessarily
correlate with the actual directory.

The filter.analyze method has been updated to accept the build output
directory as an argument and the ios module has been updated to pass a
directory whose name is based off of the build directory. The build
directory was not re-used because filter.analyze doesn't handle the
goma directory and so mb analyze would cause the args.gn file to be
re-generated with a bad value for goma_dir.

Bug: 670521
Change-Id: Ia6075d0a6cd1cbe41002107a3c93e04eab4690ef
Reviewed-on: https://chromium-review.googlesource.com/c/1426178
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>