Add downstream_spec option to bot config db
Adds the option for bot configs to specify a 'downstream_spec' entry
to use instead of loading a spec file from Chromium source. This is
similar to the earlier 'source_side_spec' option, but works properly for
builders with children, as the 'source_side_spec' approach doesn't
allow them to determine which targets they need to build and isolate.

Bug: 919161
Change-Id: I0c81228dface5f01442917c0a6155d4efe232368
Reviewed-on: https://chromium-review.googlesource.com/c/1432193
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>