[V8] Reliably fail on missing shards
Before this change, a warming was printed, but a failure was only
reached when the V8 test runner properly returned with non-zero.
In case of a test runner exception, the overall recipe would not
fail.

Now steps with missing shards will be purple and the overall run
will fail.

Bug: v8:8731
Change-Id: I19acda8fd43461b0b2c8e2726cbf866004eae6ff
Reviewed-on: https://chromium-review.googlesource.com/c/1436256
Reviewed-by: Sergiy Belozorov <sergiyb@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>