Emit FindIt Flakiness metadata.
This CL adds a step to all CQ builds that do not encounter exceptional
circumstances [tests run, they pass, or fail, or produce invalid results].

This step emits metadata afor post-processing by FindIt. Currently, the only
metadata emitted is for "Step Layer Flakiness". This refers to tests that failed
in one step, but when rerun in a subsequent step, pass. This currently applies
to 'retry with patch', but will also apply to shard-layer retries once that
feature is added.

Change-Id: I34a0b46033757c76c3fb3e7579d2652345e52c1d
Bug: 872042
Reviewed-on: https://chromium-review.googlesource.com/c/1449073
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>