Re-add a hook in ios module for WebRTC to upload performance results
WebRTC previously used the ios recipe module directly and it had to have a code path to upload Chart JSON results.
Unfortunately, the script used there is heavily deprecated and Buildbot-specific (currently broken on LUCI!). So I decided to just remove the code in [1].
Instead, WebRTC is now using only parts of the ios recipe module [2] and calling the `collect` function itself. This is a good opportunity to go further in that direction, and just add a callback to allow WebRTC to use the same perf upload function as for all other platforms.

[1]: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1446098
[2]: https://chromium-review.googlesource.com/c/chromium/tools/build/+/1443196

Bug: webrtc:10253, chromium:862134
Change-Id: Ica510d6a550e9fda52eb919835696bbfba4b721b
Reviewed-on: https://chromium-review.googlesource.com/c/1446100
Commit-Queue: Oleh Prypin <oprypin@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>