[code coverage] Fix issue that unchanged lines are being highlighted
The way coverage data is generated is that when the raw code coverage
data becomes available, a rebased code coverage data is generated based
on the added lines mapping, and if a file is not present in the added
lines mapping, all the lines in the raw coverage data are preserved.

Before this change, if a file has no added lines, then it will not be
present in the mapping, which resulted in that all the unchanged lines
are preserved.

This CL fixes the problem by making the value an empty dictionary if a
file has no added lines.

Bug: 930366
Change-Id: I36233c231ea5e97afde70e5c57bf6a6010c583c7
Reviewed-on: https://chromium-review.googlesource.com/c/1464382
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>