Add some master scripts to recipe bundle for gatekeeper
Unfortunately, these scripts are currently necessary for the
gatekeeper_ng.py script to run. I am honestly not sure if
the transitive dependencies imported by auth.py here are
going to be importable; the easiest way to tell is to submit
this CL and see if the LUCI gatekeeper builder continues to
fail or not.

Bug: 853854
Change-Id: Iad4c65cd495cf391a28ae65ea1be0baee336e1a6
Reviewed-on: https://chromium-review.googlesource.com/c/1470973
Auto-Submit: Aaron Gable <agable@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>