Implement support for custom shard dispatch of swarming tasks.
Previously, the interface for the swarming recipe allowed dispatch of all
shards, or a single shard. This CL adds support for dispatch of any arbitrary
set of shards.

To keep the new logic simple, and to prepare for the migration to the Go
implementation of "swarming.py trigger", which doesn't support the --shards
parameter, this CL dispatches individual steps for each triggered shard.

An alternative considered and discarded was to make another wrapper script that
would handle multiple-shard dispatch. This moves easily testable logic from the
recipe into a python script, which adds fragility.

One slight behavior change is that if the trigger step fails [e.g. retcode !=
0], then the links URL is not populated.

For an example of a build with this new logic using led, see:
https://ci.chromium.org/swarming/task/43222369667f5b10?server=chromium-swarm.appspot.com

Change-Id: If3560d0675ce326d2972f43d35e5ae010bae0e34
Bug: 926987
Recipe-Manual-Change: build_limited_scripts_slave
Reviewed-on: https://chromium-review.googlesource.com/c/1466191
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>