[dart] Add linearized history for multiple repositories recipe
This recipe checks out the linearized history repository specified by
the "repo" recipe property and pushes a commit to it that mimics the
gitiles_commit of the change that triggered it. A builder running this
recipe is intended to be triggered from multiple gitiles triggers and
will store the hashes from the different repositories in a commits.json
file that is updated whenever triggered.

Change-Id: Iec44557ea99718287177b6ea030a9ad316072ebf
Reviewed-on: https://chromium-review.googlesource.com/c/1491556
Commit-Queue: Alexander Thomas <athom@google.com>
Reviewed-by: Martin Kustermann <kustermann@google.com>