Do not consider missing shard results as an internal swarming failure.
A null shard in the summary indicates that the test runner did not
produce the expected output.json (usually due to the test runner not
handling a SIGTERM due to the task timing out). Identifying it as an
internal swarming failure is misleading.

Change-Id: I94811c66f8c5456831593e73604f1e7751f5faf1
Reviewed-on: https://chromium-review.googlesource.com/c/1491912
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>