Fix parsing of Blink Layout test results.
Previously, a result of TIMEOUT, PASS would be treated by the recipe as a
passing test in 'retry with patch', even though we want it to be treated as a
failing test. In general, the problem is that the chromium_test recipe was using
the method 'failures_or_invalid_results' to fetch failures, but that method was
insufficiently precise. It did not establish whether flaky tests were treated as
successes or failures. This CL updates the chromium_test recipe to use more
specific methods that explicitly handle flaky tests.

Change-Id: Ib37c4bfc3b2945c7da4ce0fa37e80d7940c38c10
Bug: 872042
Reviewed-on: https://chromium-review.googlesource.com/c/1493425
Commit-Queue: Erik Chen <erikchen@chromium.org>
Auto-Submit: Erik Chen <erikchen@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>