Remove dead priority adjustment code.
The priority adjustment config is not [correctly] used outside of tests. This is
good, because attempting to increase the priority would bump CQ tasks to
swarming priority 20, which is higher priority than the CI waterfall. That is
undesirable. In general, a blank +/- 10 bump of priority is difficult to reason
about, and isn't even guaranteed to have well defined behavior if exceeding the
accepted range for swarming priorities [0-255]. If there is a valid use case in
the future, we can implement a more robust mechanism for tests to change
priority.

Bug: 937151
Change-Id: I9e2d2664f599078b4e9f2ebf335be307fd31090d
Reviewed-on: https://chromium-review.googlesource.com/c/1495718
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>