Increase priority for 'without patch' and 'retry with patch' steps.
Increase priority of 'without patch' and 'retry with patch' steps to 29 on the
CQ, slightly higher than the 30 for 'with patch' steps.

If there are capacity issues, this will prioritize tasks for builds that are
half-complete. This will allow the CQ to more frequently finish builds, rather
than time them out halfway through, and then have to retry everything.

Bug: 937151
Change-Id: I89cb2e9e4f455f2240edc529a31c0c98f9082666
Reviewed-on: https://chromium-review.googlesource.com/c/1495732
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>